Add browser-hosted WASM simulator demo
