+++ Cargo.toml
@@ -59,6 +59,7 @@ members = [
     "examples/simulator/server",
     "examples/simulator/ui",
     "examples/simulator/ui-bindings",
+    "examples/simulator/wasm",
     # Deterministic query-engine event emitter (opt-in via `-p`).
     # Activates `quent-time/__test-clock-override`; excluded from default-members
     # so default builds keep `quent-time` zero-cost.
