+++ domains/query_engine/ui/Cargo.toml
@@ -5,6 +5,7 @@ edition.workspace =  true
 publish.workspace = true
 
 [dependencies]
+async-trait.workspace = true
 quent-analyzer = { path = "../../../crates/analyzer", features = ["ts"] }
 quent-dynamic-attributes = { path = "../../../crates/dynamic-attributes", features = ["ts"] }
 quent-query-engine-model = { path = "../model" }
