+++ examples/simulator/wasm/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "quent-simulator-wasm"
+version.workspace = true
+edition.workspace = true
+publish = false
+
+[features]
+generate = ["dep:quent-model", "dep:quent-simulator"]
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[[bin]]
+name = "generate-demo"
+path = "src/bin/generate.rs"
+required-features = ["generate"]
+
+[dependencies]
+async-trait.workspace = true
+postcard.workspace = true
+quent-events = { path = "../../../crates/events", features = ["serde"] }
+quent-model = { path = "../../../crates/model", optional = true }
+quent-query-engine-analyzer = { path = "../../../domains/query_engine/analyzer" }
+quent-query-engine-ui = { path = "../../../domains/query_engine/ui" }
+quent-simulator-analyzer = { path = "../analyzer" }
+quent-simulator-instrumentation = { path = "../instrumentation" }
+quent-simulator = { path = "../application", optional = true }
+quent-ui = { path = "../../../crates/ui" }
+serde.workspace = true
+serde_json.workspace = true
+uuid = { workspace = true, features = ["js"] }
+wasm-bindgen = "=0.2.127"
+wasm-bindgen-futures = "=0.4.77"
