+++ ui/package.json
@@ -28,6 +28,10 @@
     "ci:build": "pnpm --filter '@quent/*' --recursive build:ci && pnpm ci:build:app",
     "ci:build:app": "tsr generate && tsc && vite build",
     "deep-link": "tsx scripts/deep-link.ts",
+    "demo:wasm": "cargo build --manifest-path ../Cargo.toml -p quent-simulator-wasm --lib --target wasm32-unknown-unknown --release --locked && wasm-bindgen ../target/wasm32-unknown-unknown/release/quent_simulator_wasm.wasm --target web --out-dir generated/simulator-wasm",
+    "demo:data": "cargo run --manifest-path ../Cargo.toml -p quent-simulator-wasm --bin generate-demo --features generate --locked -- generated/simulator-demo.postcard",
+    "demo:build": "pnpm demo:wasm && pnpm demo:data && pnpm bindings && tsr generate && tsc && vite build --config vite.simulator.config.ts",
+    "demo:dev": "pnpm demo:wasm && pnpm demo:data && pnpm bindings && vite --config vite.simulator.config.ts",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "format": "prettier --write \"src/**/*.{ts,tsx,js,jsx,json,css,md}\"",
