+++ ui/tsconfig.node.json
@@ -7,6 +7,5 @@
     "allowSyntheticDefaultImports": true,
     "strict": true
   },
-  "include": ["vite.config.ts"]
+  "include": ["vite.config.ts", "vite.simulator.config.ts"]
 }
-
